# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "My Fucking Swaggah Jagahh.",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/GetBugReports": {
            "post": {
                "description": "GetBugReports",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BugReport"
                ],
                "summary": "GetBugReports",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetReports"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BugReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/SendBugReport": {
            "post": {
                "description": "Send But Report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BugReport"
                ],
                "summary": "Send Bug Report",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SendBugReport"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BugReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/getAllAtm": {
            "get": {
                "description": "Get Atm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetAtm"
                ],
                "summary": "Get Atm",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAtm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/getAllFees": {
            "get": {
                "description": "Get All Fees",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetFees"
                ],
                "summary": "Get All Fees",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FeeStructure"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/getInstiAtm": {
            "post": {
                "description": "Get Insti Atm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetAtm"
                ],
                "summary": "Get Insti Atm",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetAtmByInsti"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAtm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/getSpecifiedFee": {
            "post": {
                "description": "GetSpecifiedFee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetFees"
                ],
                "summary": "GetSpecifiedFee",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FeeStructureRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FeeStructure"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/janus/loanlist/": {
            "post": {
                "description": "/CoreAccounts/API/searchLoanList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Janus"
                ],
                "summary": "/CoreAccounts/API/searchLoanList",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.Use"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.LoanList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Errror"
                        }
                    }
                }
            }
        },
...
        "/janus/searchCID/": {
            "post": {
                "description": "CoreAccounts/API/mobile/api/v1/SearchCustomerCID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Janus"
                ],
                "summary": "CoreAccounts/API/mobile/api/v1/SearchCustomerCID",
                "parameters": [
                    {
                        "description": "SearchbyCid",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.Use"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.JsonStruct"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Errror"
                        }
                    }
                }
            }
        },
        "/updateBugreport": {
            "post": {
                "description": "Update Bug Report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BugReport"
                ],
                "summary": "Update Bug Report",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateBugreports"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BugReportResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.BugReport": {
            "type": "object",
            "properties": {
                "application_code": {
                    "type": "integer"
                },
                "cid": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "insti_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.BugReportResponses": {
            "type": "object",
            "properties": {
                "application_code": {
                    "type": "integer"
                },
                "cid": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "insti_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "retcode": {
                    "type": "integer"
                }
            }
        },
        "model.FeeStructure": {
            "type": "object",
            "properties": {
                "agent_income": {
                    "type": "number"
                },
                "agent_target_income": {
                    "type": "number"
                },
                "bancnet_income": {
                    "type": "number"
                },
                "bank_income": {
                    "type": "number"
                },
                "bank_income_flag": {
                    "type": "boolean"
                },
                "client_type": {
                    "type": "string"
                },
                "cmit_fee": {
                    "type": "number"
                },
                "created_by": {
                    "type": "integer"
                },
                "created_date": {
                    "type": "string"
                },
                "end_range": {
                    "type": "integer"
                },
                "fds_fee": {
                    "type": "number"
                },
                "fee_id": {
                    "type": "integer"
                },
                "last_updated_by": {
                    "type": "integer"
                },
                "last_updated_date": {
                    "type": "string"
                },
                "start_range": {
                    "type": "integer"
                },
                "telco_fee": {
                    "type": "number"
                },
                "total_charge": {
                    "type": "number"
                },
                "trans_type": {
                    "type": "string"
                }
            }
        },
        "model.FeeStructureRequest": {
            "type": "object",
            "properties": {
                "fee_id": {
                    "type": "integer"
                }
            }
        },
        "model.GetAtm": {
            "type": "object",
            "properties": {
                "atm_address": {
                    "type": "string"
                },
                "atm_city": {
                    "type": "string"
                },
                "atm_description": {
                    "type": "string"
                },
                "atm_id": {
                    "type": "integer"
                },
                "atm_latitude": {
                    "type": "number"
                },
                "atm_longitude": {
                    "type": "number"
                },
                "atm_type": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "created_date": {
                    "type": "string"
                },
                "inst_code": {
                    "type": "string"
                },
                "last_updated_by": {
                    "type": "integer"
                },
                "last_updated_date": {
                    "type": "string"
                }
            }
        },
        "model.GetAtmByInsti": {
            "type": "object",
            "properties": {
                "inst_code": {
                    "type": "string"
                }
            }
        },
        "model.GetReports": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.SendBugReport": {
            "type": "object",
            "properties": {
                "application_code": {
                    "type": "integer"
                },
                "cid": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "insti_code": {
                    "type": "integer"
                }
            }
        },
        "model.UpdateBugreports": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "route.Errror": {
            "type": "object",
            "properties": {
                "errormessage": {
                    "type": "string"
                }
            }
        },
...
        "route.JsonStruct": {
            "type": "object",
            "properties": {
                "accountNumber": {
                    "type": "string"
                },
                "address": {},
                "birthPlace": {
                    "type": "string"
                },
                "brCode": {
                    "type": "integer"
                },
                "business": {
                    "type": "integer"
                },
                "centerCode": {
                    "type": "integer"
                },
                "centerName": {},
                "cid": {
                    "type": "integer"
                },
                "civilStatus": {
                    "type": "integer"
                },
                "classification": {
                    "type": "integer"
                },
                "contact": {},
                "doBirth": {
                    "type": "string"
                },
                "doEntry": {
                    "type": "string"
                },
                "doRecognized": {
                    "type": "string"
                },
                "doResigned": {
                    "type": "string"
                },
                "dosri": {
                    "type": "integer"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "maidenFname": {
                    "type": "string"
                },
                "maidenLName": {
                    "type": "string"
                },
                "maidenMName": {
                    "type": "string"
                },
                "memberMaidenFName": {
                    "type": "string"
                },
                "memberMaidenLName": {
                    "type": "string"
                },
                "memberMaidenMName": {
                    "type": "string"
                },
                "middlename": {
                    "type": "string"
                },
                "reffered": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "searchName": {},
                "sex": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "subClassification": {
                    "type": "integer"
                },
                "title": {
                    "type": "integer"
                },
                "unitCode": {
                    "type": "integer"
                },
                "unitName": {}
            }
        },
...
        "route.LoanList": {
            "type": "object",
            "properties": {
                "acc": {
                    "type": "string"
                },
                "acctType": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "dateRelease": {
                    "type": "string"
                },
                "interest": {
                    "type": "number"
                },
                "oth": {
                    "type": "number"
                },
                "paidTerm": {
                    "type": "integer"
                },
                "principal": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "term": {
                    "type": "integer"
                }
            }
        },
...
        "route.Use": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 9
                }
            }
        }
    }
}